Module quack::arithmetic

Efficient modular arithmetic and polynomial evaluation.

Structs

ModularInteger

An element in the finite field with integers modulo a prime.

MontgomeryIntegermontgomery

A 64-bit finite field element in Montgomery form.

Traits

ModularArithmetic

Arithmetic operations and other properties of the modular integer field.

Functions

eval

Evaluate the univariate polynomial at x in a modular finite field.

eval_montgomerymontgomery

Evaluate the univariate polynomial at x, assuming that x and the coefficients are given in 64-bit Montgomery form using the same moduli.

eval_precomputepower_table

Evaluate the univariate polynomial at x in a 16-bit modular finite field, using precomputed power tables.

factorlibpari

Factor the univariate polynomial using 32-bit modular arithmetic, returning the roots.

Type Aliases

CoefficientVector

The coefficient vector defines a univariate polynomial where the constant term in the polynomial is the last element in the vector. The polynomial is monic, meaning the leading coefficient is one, and this is not included in the vector. The number of elements in the vector is the degree of the polynomial.