Trait quack::arithmetic::ModularArithmetic

pub trait ModularArithmetic {
    type SmallModulusType;
    type BigModulusType;

    // Required methods
    fn new(n: Self::SmallModulusType) -> Self;
    fn modulus() -> Self::SmallModulusType;
    fn modulus_big() -> Self::BigModulusType;
    fn value(&self) -> Self::SmallModulusType;
    fn add_assign(&mut self, rhs: Self);
    fn sub_assign(&mut self, rhs: Self);
    fn mul_assign(&mut self, rhs: Self);
    fn pow(self, power: Self::SmallModulusType) -> Self;
    fn neg(self) -> Self;
    fn inv(self) -> Self;

    // Provided methods
    fn add(self, rhs: Self) -> Self
       where Self: Sized { ... }
    fn sub(self, rhs: Self) -> Self
       where Self: Sized { ... }
    fn mul(self, rhs: Self) -> Self
       where Self: Sized { ... }

}

Arithmetic operations and other properties of the modular integer field.

Required Associated Types

type SmallModulusType

The smallest unsigned integer type that fits elements in the field.

type BigModulusType

The next largest unsigned integer type that fits elements in the field.

Required Methods

fn new(n: Self::SmallModulusType) -> Self

Creates a new element in the finite field.

The value of the new element is taken modulo a prime.

fn modulus() -> Self::SmallModulusType

The modulus prime of the finite field.

fn modulus_big() -> Self::BigModulusType

The modulus prime of the finite field, as a larger bit-width integer.

fn value(&self) -> Self::SmallModulusType

The integer value of the element, where 0 <= value < MODULUS.

MODULUS is the largest prime integer that fits in T.

fn add_assign(&mut self, rhs: Self)

Performs the += operation in the finite field.

fn sub_assign(&mut self, rhs: Self)

Performs the -= operation in the finite field.

fn mul_assign(&mut self, rhs: Self)

Performs the *= operation in the finite field.

fn pow(self, power: Self::SmallModulusType) -> Self

Raises the element to the power-th power in the finite field.

fn neg(self) -> Self

Performs the unary - operation in the finite field.

fn inv(self) -> Self

The modular multiplicative inverse of the element.

An element multiplied by its inverse is equal to one.

Provided Methods

fn add(self, rhs: Self) -> Selfwhere Self: Sized,

Performs the + operation in the finite field.

fn sub(self, rhs: Self) -> Selfwhere Self: Sized,

Performs the - operation in the finite field.

fn mul(self, rhs: Self) -> Selfwhere Self: Sized,

Performs the * operation in the finite field.

Object Safety

This trait is not object safe.